requirements.txt
apache-beam[gcp]==2.42.0
google-cloud-pubsub
google-cloud-storage
google-cloud

pandas==1.5.1
toml
wheel
fsspec
gcsfs
grpcio
numpy
pandas
scikit-learn
seaborn
requests
pvlib
cvxpy
Mosek
solar-data-tools